+++ src/gulsummaryxreftocsv/gulsummaryxreftocsv.cpp
@@ -56,7 +56,7 @@ namespace gulsummaryxreftocsv {
 		printf("\"coverage_id\", \"summary_id\", \"summaryset_id\"\n");
 
 		gulsummaryxref q;
-		int i = fread(&q, sizeof(q), 1, stdin);
+		size_t i = fread(&q, sizeof(q), 1, stdin);
 		while (i != 0) {
 			printf("%d, %d, %d\n", q.coverage_id, q.summary_id, q.summaryset_id);
 			i = fread(&q, sizeof(q), 1, stdin);
